requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
colorama>=0.4.6
tabulate>=0.9.0
PyJWT>=2.8.0
cryptography>=41.0.0
urllib3>=2.0.0
dnspython>=2.4.0